Update: Reorganize Controller
