+++ OOP/src/main/resources/FXML/ChangePassword.fxml
@@ -15,9 +15,9 @@
 <?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="800.0" prefWidth="1280.0" xmlns="http://javafx.com/javafx/24.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.example.oop.Control.ChangePasswordController">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="800.0" prefWidth="1280.0" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.example.oop.Control.ChangePasswordController">
    <children>
-      <Label layoutX="701.0" layoutY="160.0" text="ĐỔI MẬT KHẨU " textAlignment="CENTER">
+      <Label layoutX="701.0" layoutY="160.0" text="CHANGE PASSWORD" textAlignment="CENTER">
          <font>
             <Font size="22.0" />
          </font>
@@ -41,32 +41,32 @@
             <Insets left="8.0" />
          </opaqueInsets>
          <children>
-            <Label prefHeight="18.0" prefWidth="198.0" text="MẬT KHẨU HIỆN TẠI " textOverrun="WORD_ELLIPSIS" GridPane.halignment="RIGHT" GridPane.rowIndex="1" />
-            <TextField fx:id="tfCurrent" promptText="Nhập mật khẩu hiện tại " GridPane.columnIndex="1" GridPane.rowIndex="1" />
-            <Label prefHeight="18.0" prefWidth="214.0" text="NHẬP MẬT KHẨU MỚI " GridPane.halignment="RIGHT" GridPane.rowIndex="2" />
+            <Label prefHeight="18.0" prefWidth="198.0" text="CURRENT PASSWORD" textOverrun="WORD_ELLIPSIS" GridPane.halignment="RIGHT" GridPane.rowIndex="1" />
+            <TextField fx:id="tfCurrent" promptText="Enter..." GridPane.columnIndex="1" GridPane.rowIndex="1" />
+            <Label prefHeight="18.0" prefWidth="214.0" text="NEW PASSWORD" GridPane.halignment="RIGHT" GridPane.rowIndex="2" />
             <TextField fx:id="tfNew" promptText="Tối thiểu 8 kí tự " GridPane.columnIndex="1" GridPane.rowIndex="2" />
-            <PasswordField fx:id="pfNew" promptText="Tối thiểu 8 kí tự " GridPane.columnIndex="1" GridPane.rowIndex="2" />
-            <Label prefHeight="18.0" prefWidth="228.0" text="XÁC NHẬN MẬT KHẨU " GridPane.halignment="RIGHT" GridPane.rowIndex="4" />
+            <PasswordField fx:id="pfNew" promptText="Min 8 character" GridPane.columnIndex="1" GridPane.rowIndex="2" />
+            <Label prefHeight="18.0" prefWidth="228.0" text="CONFIRM PASSWORD" GridPane.halignment="RIGHT" GridPane.rowIndex="4" />
             <CheckBox fx:id="cbShow" mnemonicParsing="false" prefHeight="26.0" prefWidth="74.0" text="CheckBox" GridPane.columnIndex="1" GridPane.halignment="LEFT" GridPane.rowIndex="5" />
             <Label fx:id="lblError" wrapText="true" GridPane.columnIndex="1" GridPane.rowIndex="6" />
-            <Label text="ĐỘ MẠNH" GridPane.halignment="LEFT" GridPane.rowIndex="3" />
+            <Label text="STRENGTH" GridPane.halignment="LEFT" GridPane.rowIndex="3" />
             <TextField fx:id="tfConfirm" promptText="Nhập lại mật khẩu mới " GridPane.columnIndex="1" GridPane.rowIndex="4" />
-            <PasswordField fx:id="pfConfirm" promptText="Nhập lại mật khẩu mới " GridPane.columnIndex="1" GridPane.rowIndex="4" />
+            <PasswordField fx:id="pfConfirm" promptText="Confirm new password" GridPane.columnIndex="1" GridPane.rowIndex="4" />
             <HBox alignment="CENTER_LEFT" prefHeight="100.0" prefWidth="200.0" GridPane.columnIndex="1" GridPane.rowIndex="3">
                <children>
                   <Label fx:id="lblStrength" prefHeight="18.0" prefWidth="80.0" />
                   <ProgressBar fx:id="pbStrength" prefHeight="18.0" prefWidth="238.0" progress="0.0" HBox.hgrow="ALWAYS" />
                </children>
             </HBox>
-            <Hyperlink fx:id="linkForgot" alignment="TOP_LEFT" onAction="#onForgot" prefHeight="24.0" prefWidth="183.0" text="BẠN QUÊN MẬT KHẨU Ư ?" GridPane.rowIndex="5" />
+            <Hyperlink fx:id="linkForgot" alignment="TOP_LEFT" onAction="#onForgot" prefHeight="24.0" prefWidth="183.0" text="Forget Password?" GridPane.rowIndex="5" />
             <HBox alignment="CENTER_RIGHT" maxWidth="640.0" prefHeight="100.0" prefWidth="200.0" spacing="12.0" GridPane.columnIndex="1" GridPane.rowIndex="7">
                <children>
-                  <Button fx:id="btnSave" defaultButton="true" mnemonicParsing="false" onAction="#onSave" text="LƯU THAY ĐỔI " />
-                  <Button fx:id="btnCancel" cancelButton="true" mnemonicParsing="false" onAction="#onCancel" text="HỦY" />
+                  <Button fx:id="btnSave" defaultButton="true" mnemonicParsing="false" onAction="#onSave" text="Save" />
+                  <Button fx:id="btnCancel" cancelButton="true" mnemonicParsing="false" onAction="#onCancel" text="Cancel" />
                </children>
             </HBox>
-            <Label contentDisplay="RIGHT" prefHeight="18.0" prefWidth="202.0" text="TÊN ĐĂNG NHẬP" />
-            <TextField fx:id="tfAccount" promptText="Nhập tên đăng nhập" GridPane.columnIndex="1" />
+            <Label contentDisplay="RIGHT" prefHeight="18.0" prefWidth="202.0" text="ACCOUNT" />
+            <TextField fx:id="tfAccount" promptText="Enter..." GridPane.columnIndex="1" />
          </children>
       </GridPane>
    </children>
